Change DOUBLE to FLOAT for wb_term.term_weight
Needs less space, and the higher precision is not needed.

+++ repo/includes/store/sql/AddTermsWeight.sql
@@ -2,4 +2,4 @@
 -- Licence: GNU GPL v2+
 
 
-alter table /*_*/wb_terms add column term_weight DOUBLE UNSIGNED NOT NULL DEFAULT 0.0;
+alter table /*_*/wb_terms add column term_weight FLOAT UNSIGNED NOT NULL DEFAULT 0.0;
